Added ability to remove z-scoring while prserving backwards compatability
runAnalysis.py got turned into a function that can be called from metaAnalysis.py
	(Hence all the nasty white space diff.. really i didn't change much)

I added in a parameter called perNeuronVarNorm and set it to True to preserve
the behavior that had existed in the first arXiv version.

When this gets flipped to false, the default "Activity" data no longer
gets z-scored per neuron after ICA..

Instead here is what happens.. when ICA is run the values of the signal get rescaled such that their mean and variance match the raw GCaMP mean and variance. Then the (I-I0)/I0 is used, where I0 is defined to be the per-neuron 20th percentile value.

Interestingly this changes our eigenvalue analysis in S4.py