Starting on a noise model
Goal here is to build a null model whereby we assume that our observed
neural dynamics exists exclusively of locomotion signals and noise
(including noise from motion artifact). Then we will try to reject this
null model so as to argue that additional brain signals exist, which
we suspect correspond to sensory processing or long lived internal brain
states.