Dirty State. Overhauling preprcessing. keep nans, no interp, plot I
* trying to preserve all of the NaNs
* removing all smoothing and interpolation
* trying to plot intensity I directly, and not do any fold-changes

So far I modified ICA to work with nans..
I set Activity and ActivityFull to have Nans.

I figured out that Activity truncates things weirdly while ActivityFull
does not. Still not clear how that relates to the NaNs.

TODO:
Next step is to use the debugger and fix all the nan-related errors in
the PCA related subroutines.